Feat: 디자이너 리스트 뷰의 디자이너 평점 구현

+++ Backend/Chu/src/main/java/com/chu/designer/domain/Designer.java
@@ -31,7 +31,7 @@ public class Designer {
     @Embedded
     private ImagePath imagePath;
 
-    private Float reviewScore;
+    private Double reviewScore;
 
     private Integer cost;
     private LocalDateTime createdDate;
